Exploring average survival times
We have a loop over values of the edge density (increments of 0.05), and we estimate the mean survival time. The expected load is 10%. I feel that 10% may be too high, so I'm gonna decrease it.

I also need to make the code more modular so I'm not copying the loop constantly.